+++ src/main/java/com/nisovin/shopkeepers/debug/DebugOptions.java
@@ -33,6 +33,9 @@ public final class DebugOptions {
 	// Logs whenever we explicitly convert items to Spigot's data format. Note that this does not log when items get
 	// implicitly converted, which may happen under various circumstances.
 	public static final String itemConversions = add("item-conversions");
+	// Logs detailed item information for the selected trade and the items in the input slots whenever a player clicks
+	// an empty trading result slot.
+	public static final String emptyTrades = add("empty-trades");
 
 	private static String add(String debugOption) {
 		allOptions.add(debugOption);